+++ tests/resources/data/masses/c_diff_monomer_masses.csv
@@ -1,0 +1,227 @@
+﻿Structure,Monoisotopicmass
+gm(-Ac) |0,456.1956
+gm(-Ac) (x2) |0,934.3755
+gm(-Ac) (x3) |0,1412.5554
+gm(-Ac) (x4) |0,1890.7353
+gm(-Ac) (x5) |0,2368.9152
+gm(-Ac)-AE |1,656.2753
+gm(-Ac)-AEJ |1,828.3601
+gm(-Ac)-AEJA |1,899.3972
+gm(-Ac)-AEJAA |1,970.4343
+gm(-Ac)-AEJC |1,931.3738
+gm(-Ac)-AEJCA |1,1002.4109
+gm(-Ac)-AEJCC |1,1034.383
+gm(-Ac)-AEJCE |1,1059.4687
+gm(-Ac)-AEJCG |1,988.3953
+gm(-Ac)-AEJCH |1,1068.4327
+gm(-Ac)-AEJCK |1,1059.4688
+gm(-Ac)-AEJCN |1,1045.4167
+gm(-Ac)-AEJCP |1,1028.4266
+gm(-Ac)-AEJCS |1,1018.4058
+gm(-Ac)-AEJD |1,943.3918
+gm(-Ac)-AEJDA |1,1014.4289
+gm(-Ac)-AEJDC |1,1046.401
+gm(-Ac)-AEJDD |1,1058.4187
+gm(-Ac)-AEJDE |1,1071.4867
+gm(-Ac)-AEJDG |1,1000.4133
+gm(-Ac)-AEJDH |1,1080.4507
+gm(-Ac)-AEJDN |1,1057.4347
+gm(-Ac)-AEJDP |1,1040.4446
+gm(-Ac)-AEJDQ |1,1071.4504
+gm(-Ac)-AEJDT |1,1044.4395
+gm(-Ac)-AEJDW |1,1129.4711
+gm(-Ac)-AEJE |1,957.4078
+gm(-Ac)-AEJEA |1,1028.4449
+gm(-Ac)-AEJED |1,1072.4347
+gm(-Ac)-AEJEE |1,1085.5027
+gm(-Ac)-AEJEG |1,1014.4293
+gm(-Ac)-AEJEH |1,1094.4667
+gm(-Ac)-AEJEI |1,1070.4919
+gm(-Ac)-AEJEN |1,1071.4507
+gm(-Ac)-AEJES |1,1044.4398
+gm(-Ac)-AEJEV |1,1056.4762
+gm(-Ac)-AEJF |1,975.4328
+gm(-Ac)-AEJFA |1,1046.4699
+gm(-Ac)-AEJFC |1,1078.442
+gm(-Ac)-AEJFD |1,1090.4597
+gm(-Ac)-AEJFE |1,1103.5277
+gm(-Ac)-AEJFF |1,1122.5012
+gm(-Ac)-AEJFG |1,1032.4543
+gm(-Ac)-AEJFH |1,1112.4917
+gm(-Ac)-AEJFI |1,1088.5169
+gm(-Ac)-AEJFK |1,1103.5278
+gm(-Ac)-AEJFN |1,1089.4757
+gm(-Ac)-AEJFP |1,1072.4856
+gm(-Ac)-AEJFQ |1,1103.4914
+gm(-Ac)-AEJFR |1,1131.5339
+gm(-Ac)-AEJFS |1,1062.4648
+gm(-Ac)-AEJFT |1,1076.4805
+gm(-Ac)-AEJFV |1,1074.5012
+gm(-Ac)-AEJFW |1,1161.5121
+gm(-Ac)-AEJFY |1,1138.4961
+gm(-Ac)-AEJG |1,885.3815
+gm(-Ac)-AEJGA |1,956.4187
+gm(-Ac)-AEJGG |1,942.403
+gm(-Ac)-AEJGH |1,1022.4405
+gm(-Ac)-AEJGM |1,1016.422
+gm(-Ac)-AEJGP |1,982.4343
+gm(-Ac)-AEJGQ |1,1013.4401
+gm(-Ac)-AEJGR |1,1041.4827
+gm(-Ac)-AEJGS |1,972.4136
+gm(-Ac)-AEJGV |1,984.45
+gm(-Ac)-AEJGW |1,1071.4609
+gm(-Ac)-AEJH |1,965.4238
+gm(-Ac)-AEJHA |1,1036.4609
+gm(-Ac)-AEJHE |1,1093.5187
+gm(-Ac)-AEJHH |1,1102.4827
+gm(-Ac)-AEJHP |1,1062.4766
+gm(-Ac)-AEJHQ |1,1093.4824
+gm(-Ac)-AEJHS |1,1052.4558
+gm(-Ac)-AEJHT |1,1066.4715
+gm(-Ac)-AEJHW |1,1151.5031
+gm(-Ac)-AEJI |1,941.4488
+gm(-Ac)-AEJIA |1,1012.4859
+gm(-Ac)-AEJID |1,1056.4757
+gm(-Ac)-AEJIE |1,1069.5437
+gm(-Ac)-AEJIH |1,1078.5077
+gm(-Ac)-AEJII |1,1054.5329
+gm(-Ac)-AEJIK |1,1069.5438
+gm(-Ac)-AEJIN |1,1055.4917
+gm(-Ac)-AEJIP |1,1038.5016
+gm(-Ac)-AEJIQ |1,1069.5074
+gm(-Ac)-AEJIS |1,1028.4808
+gm(-Ac)-AEJIT |1,1042.4965
+gm(-Ac)-AEJIW |1,1127.5281
+gm(-Ac)-AEJK |1,956.4598
+gm(-Ac)-AEJKA |1,1027.4969
+gm(-Ac)-AEJKD |1,1071.4867
+gm(-Ac)-AEJKE |1,1084.5547
+gm(-Ac)-AEJKG |1,1013.4813
+gm(-Ac)-AEJKH |1,1093.5187
+gm(-Ac)-AEJKI |1,1069.5439
+gm(-Ac)-AEJKK |1,1084.5548
+gm(-Ac)-AEJKN |1,1070.5027
+gm(-Ac)-AEJKP |1,1053.5126
+gm(-Ac)-AEJKQ |1,1084.5184
+gm(-Ac)-AEJKR |1,1112.5562
+gm(-Ac)-AEJKT |1,1057.5075
+gm(-Ac)-AEJKW |1,1142.5391
+gm(-Ac)-AEJL |1,941.4488
+gm(-Ac)-AEJLC |1,1044.458
+gm(-Ac)-AEJLG |1,998.4703
+gm(-Ac)-AEJLV |1,1040.5172
+gm(-Ac)-AEJM |1,959.4048
+gm(-Ac)-AEJMA |1,1030.4419
+gm(-Ac)-AEJMC |1,1062.414
+gm(-Ac)-AEJMD |1,1074.4317
+gm(-Ac)-AEJME |1,1087.4997
+gm(-Ac)-AEJMF |1,1106.4732
+gm(-Ac)-AEJMH |1,1096.4637
+gm(-Ac)-AEJMI |1,1072.4889
+gm(-Ac)-AEJMK |1,1087.4998
+gm(-Ac)-AEJMM |1,1090.4453
+gm(-Ac)-AEJMN |1,1073.4477
+gm(-Ac)-AEJMP |1,1056.4576
+gm(-Ac)-AEJMQ |1,1087.4634
+gm(-Ac)-AEJMS |1,1046.4368
+gm(-Ac)-AEJMT |1,1060.4525
+gm(-Ac)-AEJMV |1,1058.4732
+gm(-Ac)-AEJMW |1,1145.4841
+gm(-Ac)-AEJMY |1,1122.4681
+gm(-Ac)-AEJNA |1,1013.4449
+gm(-Ac)-AEJNE |1,1070.5027
+gm(-Ac)-AEJNG |1,999.4293
+gm(-Ac)-AEJNH |1,1079.4667
+gm(-Ac)-AEJNN |1,1056.4507
+gm(-Ac)-AEJNP |1,1039.4606
+gm(-Ac)-AEJNQ |1,1070.4664
+gm(-Ac)-AEJNT |1,1043.4555
+gm(-Ac)-AEJNW |1,1128.4871
+gm(-Ac)-AEJP |1,925.4178
+gm(-Ac)-AEJPA |1,996.4549
+gm(-Ac)-AEJPE |1,1053.5127
+gm(-Ac)-AEJPP |1,1022.4706
+gm(-Ac)-AEJPQ |1,1053.4764
+gm(-Ac)-AEJPT |1,1026.4655
+gm(-Ac)-AEJPW |1,1111.4971
+gm(-Ac)-AEJQA |1,1027.4609
+gm(-Ac)-AEJQC |1,1059.433
+gm(-Ac)-AEJQE |1,1084.5187
+gm(-Ac)-AEJQG |1,1013.4453
+gm(-Ac)-AEJQQ |1,1084.4824
+gm(-Ac)-AEJQS |1,1043.4558
+gm(-Ac)-AEJQT |1,1057.4715
+gm(-Ac)-AEJQW |1,1142.5031
+gm(-Ac)-AEJR |1,984.4648
+gm(-Ac)-AEJRA |1,1055.5019
+gm(-Ac)-AEJRC |1,1087.474
+gm(-Ac)-AEJRD |1,1099.4917
+gm(-Ac)-AEJRE |1,1112.5597
+gm(-Ac)-AEJRH |1,1121.5237
+gm(-Ac)-AEJRI |1,1097.5489
+gm(-Ac)-AEJRM |1,1115.5053
+gm(-Ac)-AEJRN |1,1098.5077
+gm(-Ac)-AEJRP |1,1081.5176
+gm(-Ac)-AEJRQ |1,1112.5234
+gm(-Ac)-AEJRR |1,1140.5659
+gm(-Ac)-AEJRT |1,1085.5125
+gm(-Ac)-AEJRV |1,1083.5332
+gm(-Ac)-AEJRW |1,1170.5441
+gm(-Ac)-AEJS |1,915.3968
+gm(-Ac)-AEJSA |1,986.4339
+gm(-Ac)-AEJSD |1,1030.4237
+gm(-Ac)-AEJSE |1,1043.4917
+gm(-Ac)-AEJSK |1,1043.4918
+gm(-Ac)-AEJSN |1,1029.4397
+gm(-Ac)-AEJSP |1,1012.4496
+gm(-Ac)-AEJSQ |1,1043.4554
+gm(-Ac)-AEJSR |1,1071.4979
+gm(-Ac)-AEJSS |1,1002.4288
+gm(-Ac)-AEJST |1,1016.4445
+gm(-Ac)-AEJSW |1,1101.4761
+gm(-Ac)-AEJT |1,929.4128
+gm(-Ac)-AEJTA |1,1000.4499
+gm(-Ac)-AEJTC |1,1032.422
+gm(-Ac)-AEJTE |1,1057.5077
+gm(-Ac)-AEJTG |1,986.4343
+gm(-Ac)-AEJTK |1,1057.5078
+gm(-Ac)-AEJTT |1,1030.4605
+gm(-Ac)-AEJTV |1,1028.4812
+gm(-Ac)-AEJV |1,927.4328
+gm(-Ac)-AEJVA |1,998.4699
+gm(-Ac)-AEJVC |1,1030.442
+gm(-Ac)-AEJVD |1,1042.4597
+gm(-Ac)-AEJVE |1,1055.5277
+gm(-Ac)-AEJVH |1,1064.4917
+gm(-Ac)-AEJVK |1,1055.5278
+gm(-Ac)-AEJVN |1,1041.4757
+gm(-Ac)-AEJVP |1,1024.4856
+gm(-Ac)-AEJVQ |1,1055.4914
+gm(-Ac)-AEJVS |1,1014.4648
+gm(-Ac)-AEJVT |1,1028.4805
+gm(-Ac)-AEJVV |1,1026.5012
+gm(-Ac)-AEJVW |1,1113.5121
+gm(-Ac)-AEJW |1,1014.4448
+gm(-Ac)-AEJWA |1,1085.4819
+gm(-Ac)-AEJWC |1,1117.454
+gm(-Ac)-AEJWE |1,1142.5397
+gm(-Ac)-AEJWT |1,1115.4925
+gm(-Ac)-AEJWW |1,1200.5241
+gm(-Ac)-AEJY |1,991.4248
+gm(-Ac)-AEJYA |1,1062.4619
+gm(-Ac)-AEJYC |1,1094.434
+gm(-Ac)-AEJYD |1,1106.4517
+gm(-Ac)-AEJYE |1,1119.5197
+gm(-Ac)-AEJYG |1,1048.4463
+gm(-Ac)-AEJYH |1,1128.4837
+gm(-Ac)-AEJYI |1,1104.5089
+gm(-Ac)-AEJYK |1,1119.5198
+gm(-Ac)-AEJYN |1,1105.4677
+gm(-Ac)-AEJYP |1,1088.4776
+gm(-Ac)-AEJYQ |1,1119.4834
+gm(-Ac)-AEJYR |1,1147.5259
+gm(-Ac)-AEJYS |1,1078.4568
+gm(-Ac)-AEJYT |1,1092.4725
+gm(-Ac)-AEJYV |1,1090.4932
+gm(-Ac)-AEJYW |1,1177.5041
+gm(-Ac)-AEJYY |1,1154.4881
