+++ tests/resources/data/masses/e_coli_monomer_masses.csv
@@ -1,0 +1,229 @@
+﻿Structure,Monoisotopicmass
+gm|0,498.2061
+gm (x2)|0,976.386
+gm (x3)|0,1454.5659
+gm (x4)|0,1932.7458
+gm (x5)|0,2410.9257
+gm (x6)|0,2889.1056
+gm (x7)|0,3367.2855
+gm-AE|1,698.2858
+gm-AEJ|1,870.3706
+gm-AEJA|1,941.4077
+gm-AEJAA|1,1012.4448
+gm-AEJC|1,973.3843
+gm-AEJCA|1,1044.4214
+gm-AEJCC|1,1076.3935
+gm-AEJCE|1,1101.4792
+gm-AEJCG|1,1030.4058
+gm-AEJCH|1,1110.4432
+gm-AEJCK|1,1101.4793
+gm-AEJCN|1,1087.4272
+gm-AEJCP|1,1070.4371
+gm-AEJCS|1,1060.4163
+gm-AEJD|1,985.4023
+gm-AEJDA|1,1056.4394
+gm-AEJDC|1,1088.4115
+gm-AEJDD|1,1100.4292
+gm-AEJDE|1,1113.4972
+gm-AEJDG|1,1042.4238
+gm-AEJDH|1,1122.4612
+gm-AEJDN|1,1099.4452
+gm-AEJDP|1,1082.4551
+gm-AEJDQ|1,1113.4609
+gm-AEJDT|1,1086.45
+gm-AEJDW|1,1171.4816
+gm-AEJE|1,999.4183
+gm-AEJEA|1,1070.4554
+gm-AEJED|1,1114.4452
+gm-AEJEE|1,1127.5132
+gm-AEJEG|1,1056.4398
+gm-AEJEH|1,1136.4772
+gm-AEJEI|1,1112.5024
+gm-AEJEN|1,1113.4612
+gm-AEJES|1,1086.4503
+gm-AEJEV|1,1098.4867
+gm-AEJF|1,1017.4433
+gm-AEJFA|1,1088.4804
+gm-AEJFC|1,1120.4525
+gm-AEJFD|1,1132.4702
+gm-AEJFE|1,1145.5382
+gm-AEJFF|1,1164.5117
+gm-AEJFG|1,1074.4648
+gm-AEJFH|1,1154.5022
+gm-AEJFI|1,1130.5274
+gm-AEJFK|1,1145.5383
+gm-AEJFN|1,1131.4862
+gm-AEJFP|1,1114.4961
+gm-AEJFQ|1,1145.5019
+gm-AEJFR|1,1173.5444
+gm-AEJFS|1,1104.4753
+gm-AEJFT|1,1118.491
+gm-AEJFV|1,1116.5117
+gm-AEJFW|1,1203.5226
+gm-AEJFY|1,1180.5066
+gm-AEJG|1,927.392
+gm-AEJGA|1,998.4292
+gm-AEJGG|1,984.4135
+gm-AEJGH|1,1064.451
+gm-AEJgm|1,1058.4325
+gm-AEJGP|1,1024.4448
+gm-AEJGQ|1,1055.4506
+gm-AEJGR|1,1083.4932
+gm-AEJGS|1,1014.4241
+gm-AEJGV|1,1026.4605
+gm-AEJGW|1,1113.4714
+gm-AEJH|1,1007.4343
+gm-AEJHA|1,1078.4714
+gm-AEJHE|1,1135.5292
+gm-AEJHH|1,1144.4932
+gm-AEJHP|1,1104.4871
+gm-AEJHQ|1,1135.4929
+gm-AEJHS|1,1094.4663
+gm-AEJHT|1,1108.482
+gm-AEJHW|1,1193.5136
+gm-AEJI|1,983.4593
+gm-AEJIA|1,1054.4964
+gm-AEJID|1,1098.4862
+gm-AEJIE|1,1111.5542
+gm-AEJIH|1,1120.5182
+gm-AEJII|1,1096.5434
+gm-AEJIK|1,1111.5543
+gm-AEJIN|1,1097.5022
+gm-AEJIP|1,1080.5121
+gm-AEJIQ|1,1111.5179
+gm-AEJIS|1,1070.4913
+gm-AEJIT|1,1084.507
+gm-AEJIW|1,1169.5386
+gm-AEJK|1,998.4703
+gm-AEJKA|1,1069.5074
+gm-AEJKD|1,1113.4972
+gm-AEJKE|1,1126.5652
+gm-AEJKG|1,1055.4918
+gm-AEJKH|1,1135.5292
+gm-AEJKI|1,1111.5544
+gm-AEJKK|1,1126.5653
+gm-AEJKN|1,1112.5132
+gm-AEJKP|1,1095.5231
+gm-AEJKQ|1,1126.5289
+gm-AEJKR|1,1154.5667
+gm-AEJKT|1,1099.518
+gm-AEJKW|1,1184.5496
+gm-AEJL|1,983.4593
+gm-AEJLC|1,1086.4685
+gm-AEJLG|1,1040.4808
+gm-AEJLV|1,1082.5277
+gm-AEJM|1,1001.4153
+gm-AEJMA|1,1072.4524
+gm-AEJMC|1,1104.4245
+gm-AEJMD|1,1116.4422
+gm-AEJME|1,1129.5102
+gm-AEJMF|1,1148.4837
+gm-AEJMH|1,1138.4742
+gm-AEJMI|1,1114.4994
+gm-AEJMK|1,1129.5103
+gm-AEJMM|1,1132.4558
+gm-AEJMN|1,1115.4582
+gm-AEJMP|1,1098.4681
+gm-AEJMQ|1,1129.4739
+gm-AEJMS|1,1088.4473
+gm-AEJMT|1,1102.463
+gm-AEJMV|1,1100.4837
+gm-AEJMW|1,1187.4946
+gm-AEJMY|1,1164.4786
+gm-AEJNA|1,1055.4554
+gm-AEJNE|1,1112.5132
+gm-AEJNG|1,1041.4398
+gm-AEJNH|1,1121.4772
+gm-AEJNN|1,1098.4612
+gm-AEJNP|1,1081.4711
+gm-AEJNQ|1,1112.4769
+gm-AEJNT|1,1085.466
+gm-AEJNW|1,1170.4976
+gm-AEJP|1,967.4283
+gm-AEJPA|1,1038.4654
+gm-AEJPE|1,1095.5232
+gm-AEJPP|1,1064.4811
+gm-AEJPQ|1,1095.4869
+gm-AEJPT|1,1068.476
+gm-AEJPW|1,1153.5076
+gm-AEJQA|1,1069.4714
+gm-AEJQC|1,1101.4435
+gm-AEJQE|1,1126.5292
+gm-AEJQG|1,1055.4558
+gm-AEJQQ|1,1126.4929
+gm-AEJQS|1,1085.4663
+gm-AEJQT|1,1099.482
+gm-AEJQW|1,1184.5136
+gm-AEJR|1,1026.4753
+gm-AEJRA|1,1097.5124
+gm-AEJRC|1,1129.4845
+gm-AEJRD|1,1141.5022
+gm-AEJRE|1,1154.5702
+gm-AEJRH|1,1163.5342
+gm-AEJRI|1,1139.5594
+gm-AEJRM|1,1157.5158
+gm-AEJRN|1,1140.5182
+gm-AEJRP|1,1123.5281
+gm-AEJRQ|1,1154.5339
+gm-AEJRR|1,1182.5764
+gm-AEJRT|1,1127.523
+gm-AEJRV|1,1125.5437
+gm-AEJRW|1,1212.5546
+gm-AEJS|1,957.4073
+gm-AEJSA|1,1028.4444
+gm-AEJSD|1,1072.4342
+gm-AEJSE|1,1085.5022
+gm-AEJSK|1,1085.5023
+gm-AEJSN|1,1071.4502
+gm-AEJSP|1,1054.4601
+gm-AEJSQ|1,1085.4659
+gm-AEJSR|1,1113.5084
+gm-AEJSS|1,1044.4393
+gm-AEJST|1,1058.455
+gm-AEJSW|1,1143.4866
+gm-AEJT|1,971.4233
+gm-AEJTA|1,1042.4604
+gm-AEJTC|1,1074.4325
+gm-AEJTE|1,1099.5182
+gm-AEJTG|1,1028.4448
+gm-AEJTK|1,1099.5183
+gm-AEJTT|1,1072.471
+gm-AEJTV|1,1070.4917
+gm-AEJV|1,969.4433
+gm-AEJVA|1,1040.4804
+gm-AEJVC|1,1072.4525
+gm-AEJVD|1,1084.4702
+gm-AEJVE|1,1097.5382
+gm-AEJVH|1,1106.5022
+gm-AEJVK|1,1097.5383
+gm-AEJVN|1,1083.4862
+gm-AEJVP|1,1066.4961
+gm-AEJVQ|1,1097.5019
+gm-AEJVS|1,1056.4753
+gm-AEJVT|1,1070.491
+gm-AEJVV|1,1068.5117
+gm-AEJVW|1,1155.5226
+gm-AEJW|1,1056.4553
+gm-AEJWA|1,1127.4924
+gm-AEJWC|1,1159.4645
+gm-AEJWE|1,1184.5502
+gm-AEJWT|1,1157.503
+gm-AEJWW|1,1242.5346
+gm-AEJY|1,1033.4353
+gm-AEJYA|1,1104.4724
+gm-AEJYC|1,1136.4445
+gm-AEJYD|1,1148.4622
+gm-AEJYE|1,1161.5302
+gm-AEJYG|1,1090.4568
+gm-AEJYH|1,1170.4942
+gm-AEJYI|1,1146.5194
+gm-AEJYK|1,1161.5303
+gm-AEJYN|1,1147.4782
+gm-AEJYP|1,1130.4881
+gm-AEJYQ|1,1161.4939
+gm-AEJYR|1,1189.5364
+gm-AEJYS|1,1120.4673
+gm-AEJYT|1,1134.483
+gm-AEJYV|1,1132.5037
+gm-AEJYW|1,1219.5146
+gm-AEJYY|1,1196.4986
